refactor: resolve lint issues

+++ internal/clickhouse/log.go
@@ -8,6 +8,6 @@ type RequestLog struct {
 	Timestamp  time.Time
 	Method     string
 	Path       string
-	Status     uint16
-	DurationMs uint32
+	Status     int
+	DurationMs int
 }
